% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_file_readr.R
\name{fpr_file_readr}
\alias{fpr_file_readr}
\title{Fpr-file reader function}
\usage{
fpr_file_readr(fprfile)
}
\arguments{
\item{fprfile}{filename and path of the fpr file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades, pricematrixes,
 stemdiametervectors, and stemtypes.
}
\description{
Fpr-file reader function
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".fpr", recursive = TRUE, full.names= TRUE)
fpr_file_readr(fprfiles[1]) \%>\% str()
fpr_file_readr(fprfiles[2]) \%>\% str()
}
