% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_file_readr2.R
\name{fpr_file_readr2}
\alias{fpr_file_readr2}
\title{Fpr-file reader function}
\usage{
fpr_file_readr2(fprfile)
}
\arguments{
\item{fprfile}{filename and path of the fpr file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades, pricematrixes,
 stemdiametervectors, and stemtypes.
}
\description{
Fpr-file reader function
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".fpr", recursive = TRUE, full.names= TRUE)
fprtest1 <- fpr_file_readr2(fprfiles[1])
fprtest2 <- fpr_file_readr2(fprfiles[2])
}
