% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{fTZ}
\alias{fTZ}
\title{Get Time zone from the StanForD2010 countrycode variables.}
\usage{
fTZ(CountryCode = "")
}
\arguments{
\item{CountryCode}{codes according to ISO3166-1 standard}
}
\value{
Time zone as Country/City
}
\description{
Get Time zone from the StanForD2010 countrycode variables.
}
\examples{
fTZ("752")
}
