% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R
\name{getCombined.mwt}
\alias{getCombined.mwt}
\title{Get Combined Machine Work Time for all and within a SF2010 .hpr file}
\usage{
getCombined.mwt(doc)
}
\arguments{
\item{doc}{a StanFord2010 .mom xml-document}
}
\value{
a tibble
}
\description{
Get Combined Machine Work Time for all and within a SF2010 .hpr file
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"), pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_cmwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "combined_mwt"))]
doc <- xml2::read_xml(momfiles_cmwt[2])
getCombined.mwt(doc)
}
