% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqc_doc_support_functions.R
\name{getControlLogs}
\alias{getControlLogs}
\title{Fetch all control logs  from one stem entry (.hqc files)}
\usage{
getControlLogs(x)
}
\arguments{
\item{x}{a StanFord2010 .hqc stem node-tree}
}
\value{
a tibble with all control log diameters belonging to one stem
}
\description{
Fetch all control logs  from one stem entry (.hqc files)
}
\examples{
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = "hqc", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hqcfiles[1])
hqc_stems <- xml2::xml_find_all(doc, ".//d1:Stem")
getControlLogs(hqc_stems[1]) \%>\% dplyr::glimpse()
}
