% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getDelivery}
\alias{getDelivery}
\title{Delivery data from one delivery definition node}
\usage{
getDelivery(x)
}
\arguments{
\item{x}{is a node tree for one delivery definition}
}
\description{
Delivery data from one delivery definition node
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".fpr", ignore.case = TRUE,  recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
nodelist <- xml2::xml_find_all(doc, ".//d1:DeliveryDefinition")
getDelivery(nodelist[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(nodelist[1], getDelivery)
plyr::ldply(nodelist, getDelivery)
}
