% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getLoad}
\alias{getLoad}
\title{Load data from one Load  node}
\usage{
getLoad(x)
}
\arguments{
\item{x}{is a node tree for one Load}
}
\description{
Load data from one Load  node
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),
  pattern = ".fpr", ignore.case = TRUE,  recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
nodelist <- xml2::xml_find_all(doc, ".//d1:Load")
getLoad(nodelist[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(nodelist[1], getLoad)
plyr::ldply(nodelist, getLoad)
}
