% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getLoads2}
\alias{getLoads2}
\title{getLoads2 - alternative and hopefully faster altenative to getLoads()}
\usage{
getLoads2(doc)
}
\arguments{
\item{doc}{a StanFord2010 .fpr xml-document}
}
\value{
a tibble
}
\description{
getLoads2 - alternative and hopefully faster altenative to getLoads()
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".fpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
getLoads(doc)
}
