% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getLocation}
\alias{getLocation}
\title{Location data from one Location node}
\usage{
getLocation(x)
}
\arguments{
\item{x}{is a node tree for one location}
}
\description{
Location data from one Location node
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".fpr", ignore.case = TRUE,  recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
locationlist <- xml2::xml_find_all(doc, ".//d1:LocationDefinition")
getLocation(locationlist[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(locationlist[1], getLocation)
}
