% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getLocations}
\alias{getLocations}
\title{Get location data for all locations within a SF2010 .fpr file}
\usage{
getLocations(doc)
}
\arguments{
\item{doc}{a StanFord2010 .hpr xml-document}
}
\value{
a tibble
}
\description{
Get location data for all locations within a SF2010 .fpr file
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"), pattern = ".fpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
getLocations(doc)
}
