% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions.R
\name{getLogs}
\alias{getLogs}
\title{Get logsdata for all stems within a SF2010 .hpr file}
\usage{
getLogs(doc)
}
\arguments{
\item{doc}{a StanFord2010 .hpr xml-document}
}
\value{
a tibble
}
\description{
Get logsdata for all stems within a SF2010 .hpr file
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
all_logs <- getLogs(doc)
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hqc", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hqcfiles[1])
all_logs <- getLogs(doc)
}
