% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getMachineReportHeader}
\alias{getMachineReportHeader}
\title{Organize the header data of the StanForD2010 report into a tibble}
\usage{
getMachineReportHeader(doc)
}
\arguments{
\item{doc}{is an xml document}
}
\value{
a tibble
}
\description{
Organize the header data of the StanForD2010 report into a tibble
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getMachineReportHeader(doc)
}
