% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_file_readr.R
\name{getMom.all}
\alias{getMom.all}
\title{mom-file reader function}
\usage{
getMom.all(momfile)
}
\arguments{
\item{momfile}{filename and path of the mom file to read}
}
\value{
A list of data.frames:
machinereport_meta,
operators,
objects,
tracking,
CombProdDat = a table listing production data (stems, volumes, etc),
CombinedMachineWorkTime = a table listing work time,
OperatorWorkTime = a table listing work time for each operator,
IndividualMachineWorkTime
}
\description{
mom-file reader function
}
\examples{
pth <- system.file(package = "sf2010r")
momfiles <- list.files(pth,".mom$",recursive=TRUE,ignore.case=TRUE,full.names= TRUE)
momtest1 <- getMom.all(momfiles[1])
}
