% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_support_combined_mwt.R
\name{getMom.cmwt.data}
\alias{getMom.cmwt.data}
\title{Combined machine work time data from one combined machine work time node}
\usage{
getMom.cmwt.data(x)
}
\arguments{
\item{x}{is a node tree for one Combined machine work time entry}
}
\description{
Combined machine work time data from one combined machine work time node
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_cmwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "cmwt"))]
doc <- xml2::read_xml(momfiles_cmwt[3])
cmwtlist <- xml2::xml_find_all(doc, ".//d1:CombinedMachineWorkTime")
plyr::ldply(cmwtlist[1:min(length(cmwtlist), 4)], getMom.cmwt.data)
}
