% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_support_imwt.R
\name{getMom.imwt.production}
\alias{getMom.imwt.production}
\title{Indivdual machine work time production data from one individual machine work time node}
\usage{
getMom.imwt.production(x)
}
\arguments{
\item{x}{is a node tree for one Indivdual machine work time entry}
}
\description{
Indivdual machine work time production data from one individual machine work time node
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_imwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "imwt"))]
doc <- xml2::read_xml(momfiles_imwt[1])
imwtlist <- xml2::xml_find_all(doc, ".//d1:IndividualMachineWorkTime")
getMom.imwt.production(imwtlist[[3]]) \%>\% dplyr::glimpse()
plyr::ldply(imwtlist[1:min(length(imwtlist), 9)], getMom.imwt.production)
}
