% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R, R/mom_support_individual_mwt.R
\name{getMom.operatorshiftdefs}
\alias{getMom.operatorshiftdefs}
\title{get OperatorShiftDefinitions}
\usage{
getMom.operatorshiftdefs(doc)

getMom.operatorshiftdefs(doc)
}
\arguments{
\item{doc}{}
}
\value{
a tibble

a tibble
}
\description{
get OperatorShiftDefinitions

get OperatorShiftDefinitions
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".mom", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = momfiles, FUN = function(X){xml2::read_xml(X)})
str(getMom.operatorshiftdefs(docs[[2]]))
str(getMom.operatorshiftdefs(docs[[3]]))
str(getMom.operatorshiftdefs(docs[[1]]))
momfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".mom", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = momfiles, FUN = function(X){xml2::read_xml(X)})
str(getMom.operatorshiftdefs(docs[[2]]))
str(getMom.operatorshiftdefs(docs[[3]]))
str(getMom.operatorshiftdefs(docs[[1]]))
}
