% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getObjects}
\alias{getObjects}
\title{Extracting all cut object definitions from doc}
\usage{
getObjects(doc)
}
\arguments{
\item{doc}{should be an Stanford2010 xml document}
}
\value{
a tibble, number of rows = number of objects and sub objects in doc
}
\description{
Extracting all cut object definitions from doc
}
\examples{
sffiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr|.fpr|.mom|.hqc", recursive = TRUE, full.names= TRUE)
print(sffiles)
docs <- lapply(X = sffiles, FUN = function(X){xml2::read_xml(X)})
getObjects(docs[[8]]) \%>\% str()
lapply(docs[1:5], FUN = function(X){getObjects(X) \%>\% str()})
}
