% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getOperatorDef}
\alias{getOperatorDef}
\title{Operator def from operator definition nodetree}
\usage{
getOperatorDef(x)
}
\arguments{
\item{x}{is a node tree of operator definition}
}
\description{
Operator def from operator definition nodetree
}
\examples{
sffiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr|.fpr", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = sffiles, FUN = function(X){xml2::read_xml(X)})
OperatorList <- xml2::xml_find_all(docs[[1]], ".//d1:OperatorDefinition" )
getOperatorDef(OperatorList[[1]]) \%>\% str()
plyr::ldply(OperatorList, getOperatorDef )
}
