% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getOperators}
\alias{getOperators}
\title{Get the operator definitions from a StanFord2010 xml document}
\usage{
getOperators(doc)
}
\arguments{
\item{doc}{the xml document}
}
\value{
a tibble, one row for each operator
}
\description{
Get the operator definitions from a StanFord2010 xml document
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getOperators(doc) \%>\% str()
}
