% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_functions.R
\name{getPartialLoad}
\alias{getPartialLoad}
\title{Partial Load data from one Load  node}
\usage{
getPartialLoad(x)
}
\arguments{
\item{x}{is a node tree for one Load}
}
\description{
Partial Load data from one Load  node
}
\examples{
fprfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".fpr", ignore.case = TRUE,  recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(fprfiles[1])
nodelist <- xml2::xml_find_all(doc, ".//d1:Load")
getPartialLoad(nodelist[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(nodelist[1], getPartialLoad)
plyr::ldply(nodelist, getPartialLoad)
}
