% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProductDef}
\alias{getProductDef}
\title{Product def from product definition nodetree}
\usage{
getProductDef(x)
}
\arguments{
\item{x}{is a node tree of ProductDefinition}
}
\description{
Product def from product definition nodetree
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
ProductsList <- xml2::xml_find_all(doc, ".//d1:ProductDefinition" )
getProductDef(ProductsList[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(ProductsList, getProductDef )
}
