% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProductDefs}
\alias{getProductDefs}
\title{get Product definitions}
\usage{
getProductDefs(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a data.frame
}
\description{
get Product definitions
}
\examples{
sffiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr|.fpr", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = sffiles, FUN = function(X){xml2::read_xml(X)})
getProductDefs(docs[[1]]) \%>\% str()
getProductDefs(docs[[2]]) \%>\% str()
getProductDefs(docs[[3]]) \%>\% str()
getProductDefs(docs[[4]]) \%>\% str()
getProductDefs(docs[[5]]) \%>\% str()
}
