% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProductDiaDefs}
\alias{getProductDiaDefs}
\title{get Product diameter def limits}
\usage{
getProductDiaDefs(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a data.frame having diameter limits for each product
}
\description{
get Product diameter def limits
}
\examples{
sffiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr|.fpr", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = sffiles, FUN = function(X){xml2::read_xml(X)})
getProductDiaDefs(docs[[3]]) \%>\% str()
getProductDiaDefs(docs[[4]]) \%>\% str()
getProductDiaDefs(docs[[5]]) \%>\% str()

}
