% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProductMatrixes}
\alias{getProductMatrixes}
\title{Get all price matrixes}
\usage{
getProductMatrixes(doc)
}
\arguments{
\item{doc}{an StanFord2010 xml document having product definitions}
}
\value{
A list
}
\description{
Get all price matrixes
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
pms <- getProductMatrixes(doc)
}
