% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProducts2}
\alias{getProducts2}
\title{get Product definitions v2}
\usage{
getProducts2(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a data.frame
}
\description{
get Product definitions v2
}
\examples{
sffiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr|.fpr", recursive = TRUE, full.names= TRUE)
docs <- lapply(X = sffiles, FUN = function(X){xml2::read_xml(X)})
getProducts2(docs[[1]]) \%>\% str()
getProducts2(docs[[2]]) \%>\% str()
getProducts2(docs[[3]]) \%>\% str()
getProducts2(docs[[4]]) \%>\% str()
getProducts2(docs[[5]]) \%>\% str()
}
