% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getSpeciesGroupDef}
\alias{getSpeciesGroupDef}
\title{Species group def from one SpeciesGroupDefinition node}
\usage{
getSpeciesGroupDef(x)
}
\arguments{
\item{x}{is a node tree for one SpeciesGrouDefinition}
}
\value{
a tibble.
}
\description{
Species group def from one SpeciesGroupDefinition node
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
SpeciesList <- xml2::xml_find_all(doc, ".//d1:SpeciesGroupDefinition" )
getSpeciesGroupDef(SpeciesList[[1]]) \%>\% str()
species <- plyr::ldply(SpeciesList, getSpeciesGroupDef)
}
