% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getSpeciesGroupDefinitions}
\alias{getSpeciesGroupDefinitions}
\title{Get all species definitions within a SF2010 doc}
\usage{
getSpeciesGroupDefinitions(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a tibble
}
\description{
Get all species definitions within a SF2010 doc
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getSpeciesGroupDefinitions(doc)
}
