% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions.R
\name{getStemdata}
\alias{getStemdata}
\title{Stem data from one Stem node}
\usage{
getStemdata(x)
}
\arguments{
\item{x}{is a node tree for one stem}
}
\description{
Stem data from one Stem node
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[3])
stemlist <- xml2::xml_find_all(doc, ".//d1:Stem")
getStemdata(stemlist[[1]]) \%>\% dplyr::glimpse()
getStemdata(stemlist[[3]]) \%>\% dplyr::glimpse()
plyr::ldply(stemlist[1:2], getStemdata)
}
