% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions.R
\name{getStems}
\alias{getStems}
\title{Get stemdata for all stems within a SF2010 .hpr file}
\usage{
getStems(doc)
}
\arguments{
\item{doc}{a StanFord2010 .hpr xml-document}
}
\value{
a tibble
}
\description{
Get stemdata for all stems within a SF2010 .hpr file
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[3])
getStems(doc)
}
