% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions_II.R
\name{getStemsAndLogs_II}
\alias{getStemsAndLogs_II}
\title{Fetch all stems and all logs in hpr and hqc}
\usage{
getStemsAndLogs_II(doc)
}
\arguments{
\item{doc}{a StanFord2010 .hpr document}
}
\value{
a list with five data frames.
}
\description{
the function returns a list with five data frames;
stems is all the stems
stplogs is all single tree processed logs
mtplogs is all multi tree processed logs
stemgrades is the grades of all stems
stemdias is the diameter vector for the stem.
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
stl <- getStemsAndLogs_II(doc)
doc <- xml2::read_xml(hprfiles[2])
stl <- getStemsAndLogs_II(doc)
doc <- xml2::read_xml(hprfiles[3]) # MTPS and BoomPositioning
stl <- getStemsAndLogs_II(doc)
}
