% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getStemTypeDefs}
\alias{getStemTypeDefs}
\title{get stem type definitions from one SpeciesGroupDefinition}
\usage{
getStemTypeDefs(x)
}
\arguments{
\item{x}{is a node tree of SpeciesGrouDefinition}
}
\description{
get stem type definitions from one SpeciesGroupDefinition
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
SpeciesList <- xml2::xml_find_all(doc, ".//d1:SpeciesGroupDefinition" )
getStemTypeDefs(SpeciesList[[1]])
}
