% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getStemTypes}
\alias{getStemTypes}
\title{Get all stem type definitions for all SpeciesGroupDefinitions}
\usage{
getStemTypes(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a list
}
\description{
Get all stem type definitions for all SpeciesGroupDefinitions
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getStemTypes(doc)
}
