% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions_II.R
\name{getSTP_stemdiameters_II}
\alias{getSTP_stemdiameters_II}
\title{get SingleTreeProcessed tree's diameter vector. By default not present in the hpr files.}
\usage{
getSTP_stemdiameters_II(doc)
}
\arguments{
\item{doc}{a hpr document (xml)}
}
\value{
a tibble. If no diametervector is present, a message and NULL.
}
\description{
get SingleTreeProcessed tree's diameter vector. By default not present in the hpr files.
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1]) # does not have stemdiameters
getSTP_stemdiameters_II(doc)
doc <- xml2::read_xml(hprfiles[2])
getSTP_stemdiameters_II(doc)
doc <- xml2::read_xml(hprfiles[3]) # does  have stemdiameters
getSTP_stemdiameters_II(doc)
}
