% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R
\name{getTracking.data}
\alias{getTracking.data}
\title{Tracking data from mom-files#' @param x is a node tree for one Combined machine work time entry}
\usage{
getTracking.data(x)
}
\arguments{
\item{x}{a xml_node with TrackingCoordinates nodes}
}
\description{
Tracking data from mom-files#' @param x is a node tree for one Combined machine work time entry
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(momfiles[3])#'
trackinglist <- xml2::xml_find_all(doc, ".//d1:Tracking")
getTracking.data(trackinglist) \%>\% dplyr::glimpse()
plyr::ldply(trackinglist, getTracking.data)
}
