% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_file_readr.R
\name{hpr_file_readr}
\alias{hpr_file_readr}
\title{Hpr-file reader function}
\usage{
hpr_file_readr(hprfile)
}
\arguments{
\item{hprfile}{filename and path of the hpr file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades, pricematrixes,
 stemdiametervectors, and stemtypes.
}
\description{
Hpr-file reader function
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
hprtest1 <- hpr_file_readr(hprfiles[1])
hprtest2 <- hpr_file_readr(hprfiles[2])
hprtest3 <- hpr_file_readr(hprfiles[3])
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),
  pattern = ".hqc", recursive = TRUE, full.names= TRUE)
hqctest1 <- hpr_file_readr(hqcfiles[1])
hqctest2 <- hpr_file_readr(hqcfiles[2])
}
