% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqc_file_readr.R
\name{hqc_file_readr}
\alias{hqc_file_readr}
\title{HQC-file reader function}
\usage{
hqc_file_readr(hqcfile)
}
\arguments{
\item{hqcfile}{file filename and path of the hqc file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades,
 stemdiametervectors, and stemtypes.
}
\description{
HQC-file reader function
}
\examples{
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hqc", recursive = TRUE, full.names= TRUE)
 hqc_file_readr(hqcfiles[1]) \%>\% str()
 hqc_file_readr(hqcfiles[2]) \%>\% str()
}
