% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqc_file_readr.R
\name{hqcdata}
\alias{hqcdata}
\title{HQC-file reader function}
\usage{
hqcdata(hqcfile)
}
\arguments{
\item{hqc}{file filename and path of the hqc file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades, pricematrixes,
 stemdiametervectors, and stemtypes.
}
\description{
HQC-file reader function
}
\examples{
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),
pattern = ".hqc", recursive = TRUE, full.names= TRUE)
hqctest1 <- hqcdata(hqcfiles[1])
hqctest2 <- hqcdata(hqcfiles[2])
hqctest3 <- hqcdata(hqcfiles[3])
}
