% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_file_readr.R
\name{mom_file_readr}
\alias{mom_file_readr}
\title{mom-file reader function}
\usage{
mom_file_readr(momfile)
}
\arguments{
\item{momfile}{filename and path of the mom file to read}
}
\value{
A list of data.frames:
machinereport_meta,
operators,
objects,
tracking,
CombProdDat = a table listing production data (stems, volumes, etc),
CombinedMachineWorkTime = a table listing work time,
OperatorWorkTime = a table listing work time for each operator,
IndividualMachineWorkTime
}
\description{
mom-file reader function
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = ".mom$", recursive = TRUE,  ignore.case = TRUE,   full.names= TRUE)
mom_file_readr(momfiles[1]) \%>\% str()
mom_file_readr(momfiles[2]) \%>\% str()
mom_file_readr(momfiles[3]) \%>\% str()
momtest4 <- mom_file_readr(momfiles[4])
}
