% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{sf2010_type.convert}
\alias{sf2010_type.convert}
\title{Ensure correct data type for some StanForD2010 key variables}
\usage{
sf2010_type.convert(df)
}
\arguments{
\item{df}{should be a data.frame}
}
\value{
the same data frame, but certain variables set to
"correct" sf2010 data type
}
\description{
Ensure correct data type for some StanForD2010 key variables
}
\examples{
dataframe <- data.frame(MachineKey = 1:2L, ObjUserID = 1:2, LoggingFormDesc  = 1:2,
v3 = c("a","b"), v4 = c(1:2), v5 = c("1", "3"))
sf2010_type.convert(dataframe) \%>\% str()
}
