% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf10r_example.R
\name{sf2010r_example}
\alias{sf2010r_example}
\title{Get path to Stanford2010 example files in the package}
\usage{
sf2010r_example(file = NULL, fileending = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}

\item{fileending}{one of the file endings in StanFord2010; "hpr", "hqc" etc}
}
\description{
sf2010r comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access.
The function is inspired by readr::readr_example()
}
\examples{
sf2010r_example()
sf2010r_example(fileending = "hpr")
}
