% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{type_convert_sf2010}
\alias{type_convert_sf2010}
\title{Type converter for sf2010 data frames}
\usage{
type_convert_sf2010(df)
}
\arguments{
\item{df}{data.frame or tbl_df}
}
\value{
tbl_df, with same columns as input data frame.
}
\description{
This function applies readr::type_convert()
on character columns of a data frame,
except for column named MachineKey which is forced to keep the original type,
other *Key variables which are being converted to integer if they are numeric
and except non-character columns
}
\examples{
df <- data.frame(MachineKey = rep("a", 3), ObjectKey = "1", StemKey = 1:3, dbh = as.character(seq(20, 24, length = 3)))
str(type_convert_sf2010(df))
}
