% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{xml_childs_dt}
\alias{xml_childs_dt}
\title{Make a data table of node children values and corresponding attribute values}
\usage{
xml_childs_dt(y)
}
\arguments{
\item{y}{a list xml_node with some children nodes}
}
\value{
a data table
}
\description{
This function takes a xml node and return a
data.table dataframe, where the names are the child node names
and the values are the node values.
Attributes and attribute values are included.
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
x <- xml2::xml_children(doc)[2] # Get one node of doc, the first is normally a header
xml_childs_dt(x[[1]]) # The function returns a data.table
x <- xml2::xml_children(x)
xml_childs_dt(x[[27]])
 x <- xml2::xml_children(x[[27]])
xml_childs_dt(x)
x \%>\% purrr::map_dfr( ~ xml_childs_dt(.x))  # Converted to tibble
}
