% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_support_combined_mwt.R
\name{getCombined.mwt}
\alias{getCombined.mwt}
\title{Get Combined Machine Work Time for all and within a SF2010 .hpr file}
\usage{
getCombined.mwt(doc)
}
\arguments{
\item{doc}{a StanFord2010 .mom xml-document}
}
\value{
a tibble
}
\description{
Get Combined Machine Work Time for all and within a SF2010 .hpr file
}
\examples{
pth <- system.file(package = "sf2010r")
momf <- list.files(pth,".mom",ignore.case=TRUE,recursive=TRUE,full.names=TRUE)
momf_cmwt <- momf[which(stringr::str_detect(string=momf,pattern="combined_mwt"))]
doc <- xml2::read_xml(momf_cmwt[2])
getCombined.mwt(doc)
}
