% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqc_doc_get_control_stem_values.R
\name{getControlStemValues}
\alias{getControlStemValues}
\title{Fetch all control log diameters from one log entry (.hqc files)}
\usage{
getControlStemValues(doc)
}
\arguments{
\item{x}{a StanFord2010 document}
}
\value{
a list of three data frames: ControlLogDiameters, ControlLogLength, ControlLogVolumes
}
\description{
Fetch all control log diameters from one log entry (.hqc files)
}
\examples{
hqcfiles <- list.files(path =  system.file(package = "sf2010r"),
   pattern = "hqc", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hqcfiles[1])
hqc_control_stem_values <- getControlStemValues(doc)
}
