% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R
\name{getMom.cmwt.data}
\alias{getMom.cmwt.data}
\title{Combined machine work time data from one combined machine work time node}
\usage{
getMom.cmwt.data(x)
}
\arguments{
\item{x}{is a node tree for one Combined machine work time entry}
}
\description{
Combined machine work time data from one combined machine work time node
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".mom", recursive = TRUE, full.names= TRUE)
momfiles <- list.files(path = "S:/PaTversAvProsjekter/R_koder/Helmer/R_pkgs/sf2010r/sf2010r/inst/extdata",  pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_cmwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "combined_mwt"))]
doc <- xml2::read_xml(momfiles_cmwt[1])

cmwtlist <- xml2::xml_find_all(doc, ".//d1:CombinedMachineWorkTime")
getMom.cmwt.data(cmwtlist[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(cmwtlist[1:2], getMom.cmwt.data)
}
