% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R, R/mom_support_individual_mwt.R
\name{getMom.imwt.activity}
\alias{getMom.imwt.activity}
\title{Indivdual machine work time activity data from one indivdual machine work time node}
\usage{
getMom.imwt.activity(x)

getMom.imwt.activity(x)
}
\arguments{
\item{x}{is a node tree for one Indivdual machine work time entry}
}
\description{
Indivdual machine work time activity data from one indivdual machine work time node

Indivdual machine work time activity data from one indivdual machine work time node
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_imwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "individual_mwt"))]
doc <- xml2::read_xml(momfiles_imwt[2])
imwtlist <- xml2::xml_find_all(doc, ".//d1:IndividualMachineWorkTime")
getMom.imwt.activity(imwtlist[[41]]) \%>\% dplyr::glimpse()
 plyr::ldply(imwtlist[25:42], getMom.imwt.activity)
momfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
momfiles_imwt <- momfiles[which(stringr::str_detect(string = momfiles, pattern = "individual_mwt"))]
doc <- xml2::read_xml(momfiles_imwt[2])
imwtlist <- xml2::xml_find_all(doc, ".//d1:IndividualMachineWorkTime")
getMom.imwt.activity(imwtlist[[41]]) \%>\% dplyr::glimpse()
 plyr::ldply(imwtlist[25:42], getMom.imwt.activity)
}
