% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getObjectDefinition}
\alias{getObjectDefinition}
\title{Extracting all data defining one cut object from one ObjectDefinition xml-tree}
\usage{
getObjectDefinition(x)
}
\arguments{
\item{x}{the object definition XML-tree}
}
\value{
a tibble, number of rows = number of sub objects
}
\description{
Extracting all data defining one cut object from one ObjectDefinition xml-tree
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
Objects_nodes <- xml2::xml_find_all(doc, "//d1:ObjectDefinition")
getObjectDefinition(Objects_nodes[1])
plyr::ldply(Objects_nodes, getObjectDefinition)
}
