% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getObjects}
\alias{getObjects}
\title{Extracting all cut object definitions from doc}
\usage{
getObjects(doc)
}
\arguments{
\item{doc}{should be an Stanford2010 xml document}
}
\value{
a tibble, number of rows = number of objects and sub objects in doc
}
\description{
Extracting all cut object definitions from doc
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getObjects(doc) \%>\% str()
}
