% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getOperatorDef}
\alias{getOperatorDef}
\title{Operator def from operator definition nodetree}
\usage{
getOperatorDef(x)
}
\arguments{
\item{x}{is a node tree of operator definition}
}
\description{
Operator def from operator definition nodetree
}
\examples{
hprfiles <-  list.files(path =  system.file(package = "sf2010r"),
   pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
OperatorList <- xml2::xml_find_all(doc, ".//d1:OperatorDefinition" )
getOperatorDef(OperatorList[[1]]) \%>\% dplyr::glimpse()
plyr::ldply(OperatorList, getOperatorDef )
}
