% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2010_general_functions.R
\name{getProductDefs}
\alias{getProductDefs}
\title{Get all product definitions}
\usage{
getProductDefs(doc)
}
\arguments{
\item{doc}{a StanFord2010 xml document}
}
\value{
a list
}
\description{
Get all product definitions
}
\examples{
pth <- system.file(package = "sf2010r")
hprfiles <-  list.files(pth,".hpr",recursive=TRUE,full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getProductDefs(doc)
doc <- xml2::read_xml(hprfiles[2])
getProductDefs(doc)
}
