% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions.R
\name{getStemsAndLogs}
\alias{getStemsAndLogs}
\title{Fetch all stems and all logs in hpr}
\usage{
getStemsAndLogs(doc)
}
\arguments{
\item{doc}{a StanFord2010 .hpr document}
}
\value{
a list
}
\description{
Fetch all stems and all logs in hpr
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"), pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
stl <- getStemsAndLogs(doc)
doc <- xml2::read_xml(hprfiles[2])
stl <- getStemsAndLogs(doc)
doc <- xml2::read_xml(hprfiles[3])
stl <- getStemsAndLogs(doc)

}
