% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_doc_support_functions.R
\name{getSTP_stemdiameters}
\alias{getSTP_stemdiameters}
\title{get SingleTreeProcessed tree's diametres}
\usage{
getSTP_stemdiameters(doc)
}
\arguments{
\item{doc}{a hpr document (xml)}
}
\value{
a tibble. If no diametervector is present, a message and NULL.
}
\description{
get SingleTreeProcessed tree's diametres
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"), pattern = ".hpr", recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(hprfiles[1])
getSTP_stemdiameters(doc)
doc <- xml2::read_xml(hprfiles[2])
getSTP_stemdiameters(doc)
doc <- xml2::read_xml(hprfiles[3])
getSTP_stemdiameters(doc)
}
