% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_functions.R
\name{getTracking.data}
\alias{getTracking.data}
\title{Tracking data from mom-files#'}
\usage{
getTracking.data(doc)
}
\arguments{
\item{doc}{a StanFord2010 .mom xml-document}
}
\description{
Tracking data from mom-files#'
}
\examples{
momfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".mom", ignore.case = TRUE, recursive = TRUE, full.names= TRUE)
doc <- xml2::read_xml(momfiles[3])#'
getTracking.data(doc) \%>\% dplyr::glimpse()
}
