% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_file_readr.R
\name{hpr_file_readr}
\alias{hpr_file_readr}
\title{Hpr-file reader function}
\usage{
hpr_file_readr(hprfile, read.diavector = FALSE)
}
\arguments{
\item{hprfile}{filename and path of the hpr file to read}

\item{read.diavector}{flag to read diameter vector or not.
if read.diavector is TRUE, function may take long time.}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta,
 operators, objects, stem_grades, pricematrixes,
 stemdiametervectors, and stemtypes.
}
\description{
Hpr-file reader function
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),
  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
hpr_file_readr(hprfiles[1], read.diavector = TRUE)
hpr_file_readr(hprfiles[2]) \%>\% str()
hpr_file_readr(hprfiles[3]) \%>\% str()
hpr_file_readr(hprfiles[3], read.diavector = TRUE) \%>\% str()
}
