% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr_file_readr.R
\name{hprdata}
\alias{hprdata}
\title{Hpr-file reader function}
\usage{
hprdata(hprfile)
}
\arguments{
\item{hprfile}{filename and path of the hpr file to read}
}
\value{
A list of data.frames: stems, products, logs, machinereport_meta, operators, objects, stem_grades, pricematrixes, stemdiametervectors, and stemtypes.
}
\description{
Hpr-file reader function
}
\examples{
hprfiles <- list.files(path =  system.file(package = "sf2010r"),  pattern = ".hpr", recursive = TRUE, full.names= TRUE)
hprtest1 <- hprdata(hprfiles[1])
hprtest2 <- hprdata(hprfiles[2])
hprtest3 <- hprdata(hprfiles[3])
hprtests <- purrr::map_dfr(hprfiles, ~tibble::tibble(filename = .x, hprdata = list(hprdata(.x))))
}
